requirements.txt
pymongo
python-dotenv
google-cloud-storage
requests
ffmpeg-python
loguru
schedule 
pytest 
pydantic 
feedparser
beautifulsoup4